Revert "Implemented search feature for outlet via filter route"
