+++ api/src/app/routes/outletRoutes.py
@@ -1,26 +1,14 @@
-from appTypes.outletTypes import (
-    NewFoodOutletBodyParams,
-    UpdateFoodOutletBodyParams,
-    GetAllFoodOutletDetailsResponseModel,
-    GetFoodOutletDetailsResponseModel,
-    CreateFoodOutletResponseModel,
-    UpdateFoodOutletResponseModel,
-    FilterFoodOutletsResponseModel,
-    FilterFoodOutletBodyParams,
-)
+from appTypes.outletTypes import FoodOutletDetails, NewFoodOutletBodyParams
 from app.app import app
-from app.db import connect, disconnect
-from app.interfaces.outlet import FoodOutlet, searchOutlets
-from app.interfaces.common import Location, Rating
+from db import connect, disconnect
+from app.outlet.outlet import FoodOutlet
 from app.auth.key import get_api_key
 from fastapi import Depends, status, HTTPException
 from fastapi.security.api_key import APIKey
 
 
 @app.get("/food-outlet")
-async def get_all_food_outlet_details(
-    api_key: APIKey = Depends(get_api_key),
-) -> GetAllFoodOutletDetailsResponseModel:
+async def get_all_food_outlet_details(api_key: APIKey = Depends(get_api_key)):
     con = connect()
     cursor = con.cursor()
 
@@ -31,70 +19,35 @@ async def get_all_food_outlet_details(
 
     for id in ids:
         outlet = FoodOutlet(id=id)
-        await outlet.sync_details(con)
-        outlet_details = outlet.__dict__
-        outlet_details["location"] = outlet.location.__dict__
-        outlet_details["rating"] = outlet.rating.__dict__
-        details.append(outlet_details)
+        outlet.sync_details(con)
+        details.append(outlet.__dict__)
 
     disconnect(con)
-    return {"outlets": details}
+    return details
 
 
 @app.get("/food-outlet/{id}")
 async def get_food_outlet_details(
     id: int, api_key: APIKey = Depends(get_api_key)
-) -> GetFoodOutletDetailsResponseModel:
+) -> FoodOutletDetails:
     con = connect()
 
     outlet = FoodOutlet(id=id)
-    await outlet.sync_details(con)
-
-    details = outlet.__dict__
-    details["location"] = outlet.location.__dict__
-    details["rating"] = outlet.rating.__dict__
+    outlet.sync_details(con)
 
     disconnect(con)
 
-    return {"outlet": details}
-
-
-@app.get("/food-outlet/filter")
-async def filter_food_outlets(
-    params: FilterFoodOutletBodyParams,
-    api_key: APIKey = Depends(get_api_key),
-) -> FilterFoodOutletsResponseModel:
-    con = connect()
-    outlets = await searchOutlets(
-        con=con,
-        nameFilter=params.name,
-        locationFilter=params.location,
-        landmarkFilter=params.landmark,
-        timeFilter=params.current_time,
-        ratingFilter=params.rating,
-    )
-
-    for i in range(len(outlets)):
-        outlets[i] = outlets[i].__dict__
-        outlets[i]["location"] = outlets[i]["location"].__dict__
-        outlets[i]["rating"] = outlets[i]["rating"].__dict__
-
-    return {"outlets": outlets}
+    return outlet.__dict__
 
 
 @app.post("/food-outlet", status_code=status.HTTP_201_CREATED)
 async def create_food_outlet(
     params: NewFoodOutletBodyParams, api_key: APIKey = Depends(get_api_key)
-) -> CreateFoodOutletResponseModel:
+):
     con = connect()
-    params_dict = params.model_dump()
-    if params_dict["location"] is not None:
-        params_dict["location"] = Location(**params_dict["location"])
-    if params_dict["rating"] is not None:
-        params_dict["rating"] = Rating(**params_dict["rating"])
-    outlet = FoodOutlet(**params_dict)
+    outlet = FoodOutlet(**params.model_dump())
     try:
-        await outlet.sync_details(con)
+        outlet.sync_details(con)
         disconnect(con)
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
@@ -104,42 +57,20 @@ async def create_food_outlet(
         if e.status_code != status.HTTP_404_NOT_FOUND:
             raise e
 
-    details = await outlet.create(con)
+    details = outlet.create(con)
 
     disconnect(con)
-    return {"outlet": details}
+    return details
 
-
-@app.put("/food-outlet/{id}")
-async def update_food_outlet(
-    id: int, params: UpdateFoodOutletBodyParams, api_key: APIKey = Depends(get_api_key)
-) -> UpdateFoodOutletResponseModel:
-    con = connect()
-    outlet = FoodOutlet(id=id)
-    await outlet.sync_details(con=con)
-    params_dict = params.model_dump()
-    for key, value in params_dict.items():
-        if value is not None:
-            if key == "location":
-                outlet.location = Location(**params_dict["location"])
-            elif key == "rating":
-                outlet.rating = Rating(**params_dict["rating"])
-            else:
-                outlet.__setattr__(key, value)
-
-    details = await outlet.update(con=con)
-    return {"outlet": details}
-
-
-@app.delete("/food-outlet/{id}", status_code=status.HTTP_204_NO_CONTENT)
+@app.delete("/food-outlet", status_code=status.HTTP_204_NO_CONTENT)
 async def delete_food_outlet(id: int, api_key: APIKey = Depends(get_api_key)):
     con = connect()
     outlet = FoodOutlet(id=id)
 
     try:
-        await outlet.sync_details(con)
+        outlet.sync_details(con)
     except HTTPException as e:
         if e.status_code == status.HTTP_404_NOT_FOUND:
             raise e
 
-    await outlet.remove(con)
+    outlet.remove(con)
